all: fix many compilation warnings (refs #45)

+++ src/cgame/cg_main.c
@@ -51,7 +51,7 @@ vmMain
 This is the only way control passes into the module.
 ================
 */
-Q_EXPORT int vmMain(int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11) {
+Q_EXPORT intptr_t vmMain(intptr_t command, intptr_t arg0, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4, intptr_t arg5, intptr_t arg6, intptr_t arg7, intptr_t arg8, intptr_t arg9, intptr_t arg10, intptr_t arg11) {
 	// Nico, silent GCC
 	(void)arg3;
 	(void)arg4;
