+++ src/cgame/cg_syscalls.c
@@ -29,9 +29,9 @@ If you have questions concerning this license or the applicable additional terms
 // cg_syscalls.c -- this file is only included when building a dll
 #include "cg_local.h"
 
-static int (QDECL *syscall)(int arg, ...) = (int (QDECL *)(int, ...)) - 1;
+static intptr_t (QDECL *syscall)(intptr_t arg, ...) = (intptr_t (QDECL *)(intptr_t, ...)) - 1;
 
-Q_EXPORT void dllEntry(int (QDECL *syscallptr)(int arg, ...)) {
+Q_EXPORT void dllEntry(intptr_t (QDECL *syscallptr)(intptr_t arg, ...)) {
 	syscall = syscallptr;
 }
