+++ src/game/bg_animation.c
@@ -1427,7 +1427,7 @@ int BG_GetConditionValue(int client, int condition, qboolean checkConversion) {
 			return 0;
 		}
 		// xkan, 1/14/2003 - must use COM_BitCheck on the result.
-		return (int)globalScriptData->clientConditions[client][condition];
+		return (intptr_t)globalScriptData->clientConditions[client][condition];
 	}
 	return globalScriptData->clientConditions[client][condition][0];
 }
